requirements.txt
boto3==1.17.61
botocore==1.20.61
certifi==2020.12.5
chardet==4.0.0
idna==2.10
jmespath==0.10.0
psycopg2-binary~=2.9.3
python-dateutil==2.8.1
requests==2.25.1
s3transfer==0.4.2
schedule==1.1.0
six==1.15.0
urllib3==1.26.4

bs4~=0.0.1
beautifulsoup4~=4.9.1
selenium~=4.1.3
numpy~=1.19.5
pandas~=1.2.1
matplotlib~=3.4.2
redis~=4.3.3